Drop mobile support and remove unused pages/icons
This page needs to be simple and easy to update, the most important
thing in this page is be able to generate the pdf.

Remove projects page, this projects can be found in youtube/github, no
need to explain every single one of them.

Icons are saved on repository, this way the user doesn't have to request
from other page the images (sometimes making the image load after the
page).

Style: No underline in links